+++ apps/docs/components/doc/preview/codePreview.tsx
@@ -24,6 +24,7 @@ export const CodePreview: FC<Props> = (props) => {
         customSetup={{
           dependencies: {
             "@orderly.network/hooks": "latest",
+
             axios: "latest",
             ...dependencies,
           },