+++ apps/docs/example/wallet_modal/walletButton.example.tsx
@@ -1,0 +1,14 @@
+import { useWalletConnector } from "./walletConnectProvider";
+
+export const WalletButtonExample = () => {
+  const wallet = useWalletConnector();
+  return (
+    <button
+      onClick={() => {
+        wallet.connect();
+      }}
+    >
+      Wallet connect
+    </button>
+  );
+};