+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @orderly.network/hooks
 
+## 1.0.2
+
+### Patch Changes
+
+- export usePrivateInfiniteQuery
+- Updated dependencies
+  - @orderly.network/types@0.1.1
+  - @orderly.network/core@0.1.1
+  - @orderly.network/perp@1.0.1
+  - @orderly.network/net@1.0.88
+
 ## 1.0.1
 
 ### Patch Changes